requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.4
pydantic-settings==2.3.0
python-dotenv==1.0.1
authlib==1.3.1
itsdangerous==2.2.0
google-cloud-storage==2.17.0
google-cloud-tasks==2.14.0
google-cloud-aiplatform==1.59.0
pdfplumber==0.11.0
langchain-text-splitters==0.2.0
pgvector==0.4.2
psycopg2-binary==2.9.9
pystray==0.19.5
pillow==10.4.0
watchdog==4.0.0
pytest==8.2.2
pytest-asyncio==0.23.7
ruff==0.5.0
